third_party/blink/web_tests/external/wpt/client-hints/service-workers/service-worker-new-request.https.html

<!DOCTYPE html>
<title>Simple test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test((t) => {
  navigator.serviceWorker.register('resources/new-request.js')
  .then((e) =>{
    var popup_window = window.open("resources/foo.html");
    assert_not_equals(popup_window, null, "Popup windows not allowed?");
    popup_window.addEventListener('load', (e) => {
      t.step(()=>{assert_equals(popup_window.document.body.textContent, "FAIL")});
      t.done();
    });
  })
}, "Client Hint/Service Worker interaction: service worker does NOT generate client hints in a new request");
</script>
